refactor(OwnedItems): 컴포넌트를 별도의 파일로 분리
