+++ src/pages/Character/components/StoredItems/ItemModal.styles.ts
