+++ src/pages/Character/components/StoredItems/ItemModal.tsx
@@ -2,7 +2,7 @@ import { Typography } from '@/components/common/Typography';
 import { BASE_URL } from '@/constants/routes';
 import type { StoreItem } from '../../types/Item';
 import { ItemImage } from '../ItemGrid.styles';
-import { ModalContent, PurchaseButton } from '../ItemModal.styles';
+import { ModalContent, PurchaseButton } from './ItemModal.styles';
 import { ItemHeart } from './StoredItems.styles';
 
 function ItemModal({
