+++ src/pages/Character/components/character/CharacterTab.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { TABS } from '../../constants/tab';
 import type { SelectedItem, StoreItem } from '../../types/Item';
 import type { Tab } from '../../types/tab';
-import ItemOwnedGrid from '../ItemOwnedGrid';
+import OwnedItemsGrid from '../OwnedItems/OwnedItemsGrid';
 import ItemStoreGrid from '../StoredItems/StoredItems';
 import Tabs from '../Tab';
 import { ContentContainer } from './Character.styles';
@@ -26,7 +26,7 @@ function CharacterTab({
       {tab === TABS.STORE ? (
         <ItemStoreGrid items={storeItems} />
       ) : (
-        <ItemOwnedGrid items={ownedItems} />
+        <OwnedItemsGrid items={ownedItems} />
       )}
     </ContentContainer>
   );
